enable slice on last dim, clean and rename test file
